requirements.txt
Flask==2.3.3
flask-cors==3.0.10
google-cloud-tasks==2.12.0